refactor: separate logic of google-font-store
