+++ packages/core/src/web/app/stores/googleFontStore/utils/variants.ts
@@ -1,0 +1,91 @@
+import { match, P } from 'ts-pattern';
+
+import type { GoogleFontFiles } from '@core/helpers/fonts/googleFontsApiCache';
+
+import { WEIGHT_FALLBACK_ORDER } from '../constants';
+
+export const getWeightAndStyle = (variant: string) =>
+  match(variant)
+    .with('regular', () => ({ style: 'normal', weight: 400 }) as const)
+    .with('italic', () => ({ style: 'italic', weight: 400 }) as const)
+    .with(
+      P.string.endsWith('italic'),
+      (v) => ({ style: 'italic', weight: Number.parseInt(v.replace('italic', ''), 10) }) as const,
+    )
+    .with(P.string.regex(/^\d+$/), (v) => ({ style: 'normal', weight: Number.parseInt(v, 10) }) as const)
+    .otherwise(() => ({ style: 'normal', weight: 400 }) as const);
+
+export const getCSSWeight = (variant: string) =>
+  match(variant)
+    .with('regular', () => '400')
+    .with('italic', () => '400:ital')
+    .with(P.string.endsWith('italic'), (v) => v.replace('italic', ':ital'))
+    .otherwise((variant) => variant);
+
+export const discoverAvailableVariants = (variants: string[]) => {
+  const available = new Set<keyof GoogleFontFiles>();
+
+  for (const variant of variants) {
+    match(variant)
+      .with('regular', () => {
+        available.add('regular');
+        available.add('400');
+      })
+      .with('italic', () => {
+        available.add('italic');
+        available.add('400italic');
+      })
+      .with(P.string.endsWith('italic'), () => {
+        const weight = variant.replace('italic', '');
+
+        available.add(`${weight}italic` as keyof GoogleFontFiles);
+      })
+      .otherwise(() => {
+        available.add(variant as keyof GoogleFontFiles);
+      });
+  }
+
+  return available;
+};
+
+export const findBestVariant = (
+  availableVariants: Set<keyof GoogleFontFiles>,
+  requestedWeight: number,
+  requestedStyle: 'italic' | 'normal',
+): keyof GoogleFontFiles | null => {
+  const createVariantKey = (style: 'italic' | 'normal', weight: number): keyof GoogleFontFiles =>
+    match({ style, weight })
+      .with({ style: 'normal', weight: 400 }, () => 'regular')
+      .with({ style: 'italic', weight: 400 }, () => 'italic')
+      .with({ style: 'italic' }, ({ weight }) => `${weight}italic`)
+      .with({ style: 'normal' }, ({ weight }) => `${weight}`)
+      .exhaustive() as keyof GoogleFontFiles;
+  const exactKey = createVariantKey(requestedStyle, requestedWeight);
+
+  if (availableVariants.has(exactKey)) {
+    return exactKey;
+  }
+
+  for (const weight of WEIGHT_FALLBACK_ORDER) {
+    const key = createVariantKey(requestedStyle, weight);
+
+    if (availableVariants.has(key)) {
+      return key;
+    }
+  }
+
+  const alternateStyle = requestedStyle === 'normal' ? 'italic' : 'normal';
+
+  // fallback to alternate style with same weight first
+  for (const weight of WEIGHT_FALLBACK_ORDER) {
+    const key = createVariantKey(alternateStyle, weight);
+
+    if (availableVariants.has(key)) {
+      console.warn(`Font variant fallback: ${requestedStyle} ${requestedWeight} → ${alternateStyle} ${weight}`);
+
+      return key;
+    }
+  }
+
+  return null;
+};
